refactor: Add new shared domain errors
