+++ src/shared/domain/errors/not_enough_permissions_error.go
@@ -1,0 +1,18 @@
+package errors
+
+import (
+	"fmt"
+	"net/http"
+)
+
+type NotEnoughPermissionsError struct {
+	Message string
+}
+
+func (err NotEnoughPermissionsError) Error() string {
+	return fmt.Sprintf("Not enough permissions: %s", err.Message)
+}
+
+func (err NotEnoughPermissionsError) StatusCode() int {
+	return http.StatusForbidden
+}
